+++ packages/checkbox-group/stories/chechbox-group.stories.tsx
@@ -1,6 +1,7 @@
 import * as React from "react"
 import { CheckBoxGroup } from "../src"
 import { CheckBox } from "@jfront/ui-checkbox"
+import { useState, useEffect } from "react"
 
 export default {
   title: "CheckBoxGroup",
@@ -8,23 +9,24 @@ export default {
 }
 
 export const BasicUsage = () => {
-  const onChange = () => {
-    console.log("onChange()")
-  }
+  const [values, setValues] = useState<string[]>()
 
   return (
-    <>
+    <div>
       <CheckBoxGroup
         name="name"
         text="Group"
         isLoading={false}
         disabled={false}
-        value={["asd", "as2"]}
+        values={values ? values : []}
+        onChange={(name, _values) => {
+          setValues(_values)
+        }}
       >
-        <CheckBox label="label1" value="value" />
-        {/*<CheckBox label="label2" value="value"/>*/}
+        <CheckBox label="label1" value="value1" />
+        <CheckBox label="label2" value="value2" />
       </CheckBoxGroup>
-    </>
+    </div>
   )
 }
 
@@ -35,26 +37,20 @@ export const Disabled = () => {
 
   return (
     <>
-      <CheckBoxGroup
-        name="name"
-        text="Group"
-        isLoading={false}
-        value={["asd", "as2"]}
-      >
-        <CheckBox disabled={true} label="label1" value="value" />
-        <CheckBox label="label2" value="value" />
-        <CheckBox label="label3" value="value" />
+      <CheckBoxGroup name="name" text="Group" isLoading={false}>
+        <CheckBox disabled={true} label="label1" value="value1" />
+        <CheckBox label="label2" value="value2" />
+        <CheckBox label="label3" value="value3" />
       </CheckBoxGroup>
       <CheckBoxGroup
         name="name2"
         text="DisabledGroup"
         isLoading={false}
         disabled={true}
-        value={["asd", "as2"]}
       >
-        <CheckBox label="label1" value="value" />
-        <CheckBox label="label2" value="value" />
-        <CheckBox label="label3" value="value" />
+        <CheckBox label="label1" value="value1" />
+        <CheckBox label="label2" value="value2" />
+        <CheckBox label="label3" value="value3" />
       </CheckBoxGroup>
     </>
   )
@@ -67,15 +63,9 @@ export const Loading = () => {
 
   return (
     <>
-      <CheckBoxGroup
-        name="name"
-        text="Group"
-        disabled={false}
-        value={["asd", "as2"]}
-        isLoading
-      >
-        <CheckBox label="label1" value="value" />
-        <CheckBox label="label2" value="value" />
+      <CheckBoxGroup name="name" text="Group" disabled={false} isLoading>
+        <CheckBox label="label1" value="value1" />
+        <CheckBox label="label2" value="value2" />
       </CheckBoxGroup>
     </>
   )
@@ -92,11 +82,10 @@ export const Error = () => {
         name="name"
         text="Group"
         disabled={false}
-        value={["asd", "as2"]}
         error="Wrong value"
       >
-        <CheckBox label="label1" value="value" />
-        <CheckBox label="label2" value="value" />
+        <CheckBox label="label1" value="value1" />
+        <CheckBox label="label2" value="value2" />
       </CheckBoxGroup>
     </>
   )