+++ packages/input/src/SelectInput.tsx
@@ -71,7 +71,7 @@ export const SelectInput = React.forwardRef<
   }
 
   return (
-    <div>
+    <div style={{ display: props.label ? "block" : "inline-block" }}>
       {props.label !== undefined && (
         <Label htmlFor={props.id}>{props.label}:&nbsp;</Label>
       )}