+++ packages/input/src/TextInput.tsx
@@ -13,7 +13,7 @@ export interface TextInputProps
 export const TextInput = React.forwardRef<HTMLInputElement, TextInputProps>(
   (props, ref) => {
     return (
-      <div>
+      <div style={{ display: props.label ? "block" : "inline-block" }}>
         {props.label !== undefined && (
           <Label htmlFor={props.id}>{props.label}:&nbsp;</Label>
         )}